Ghosts can see SM Stats [QoL] [PORT] (#300)
* Commit

* supermatter
