+++ code/modules/modular_computers/file_system/programs/sm_monitor.dm
@@ -115,7 +115,7 @@
 	if(computer.active_program == src)
 		computer.alert_call(src, "Crystal delamination in progress!")
 
-/datum/computer_file/program/supermatter_monitor/ui_data()
+/datum/computer_file/program/supermatter_monitor/ui_data(mob/user)
 	var/list/data = get_header_data()
 
 	if(istype(active))
@@ -128,32 +128,8 @@
 		if(!air)
 			active = null
 			return
-
-		data["active"] = TRUE
-		data["SM_integrity"] = active.get_integrity()
-		data["SM_power"] = active.power
-		data["SM_radiation"] = active.last_rads
-		data["SM_ambienttemp"] = air.return_temperature()
-		data["SM_ambientpressure"] = air.return_pressure()
-		data["SM_bad_moles_amount"] = MOLE_PENALTY_THRESHOLD / active.gasefficency
-		data["SM_moles"] = 0
-
-		var/list/gasdata = list()
-
-		if(air.total_moles())
-			data["SM_moles"] = air.total_moles()
-			for(var/gasid in air.get_gases())
-				gasdata.Add(list(list(
-				"name"= GLOB.gas_data.names[gasid],
-				"amount" = round(100*air.get_moles(gasid)/air.total_moles(),0.01))))
-
-		else
-			for(var/gasid in air.get_gases())
-				gasdata.Add(list(list(
-					"name"= GLOB.gas_data.names[gasid],
-					"amount" = 0)))
-
-		data["gases"] = gasdata
+		data += active.ui_data()
+		data["standalone_mode"] = FALSE
 	else
 		var/list/SMS = list()
 		for(var/obj/machinery/power/supermatter_crystal/S in supermatters)
@@ -161,7 +137,7 @@
 			if(A)
 				SMS.Add(list(list(
 				"area_name" = A.name,
-				"integrity" = S.get_integrity(),
+				"integrity" = S.get_integrity_percent(),
 				"uid" = S.uid
 				)))
 
@@ -171,8 +147,9 @@
 	return data
 
 /datum/computer_file/program/supermatter_monitor/ui_act(action, params)
-	if(..())
-		return TRUE
+	. = ..()
+	if(.)
+		return
 
 	switch(action)
 		if("PRG_clear")
